fix: Fixed Notifications timing/Call for prayer #47
Added a control statement that checks if time setup to remind the client of the prayer is up, so it can push a notification that notifies the client of how much time left. And edited the control statement that checks if the prayer time is up, so it tells the client that its prayer time and calls for prayer.